backend/requirements.txt
# FastAPI and server dependencies
fastapi
uvicorn[standard]
websockets
python-dotenv
pydantic<2.0.0
requests

# Browser automation
browser-use
playwright

# AI and ML dependencies
langchain-core<0.1.0
langchain-google-genai<0.1.0
google-ai-generativelanguage<0.4.0
google-api-core<3.0.0
openai

# Azure dependencies
azure-search-documents
azure-core

# Additional utilities
typing-extensions
aiohttp
python-multipart
starlette
anyio
click
h11
httptools
python-jose[cryptography]
uvloop; sys_platform != "win32"
watchfiles